Create directories pgn_process and tablebase
