+++ include/tablebase/persistence.hpp
@@ -1,0 +1,6 @@
+#include <string>
+
+class TablebasePersistor
+{
+    void serialize_tablebase(std::string file_path)
+}
