+++ src/process_pgn/pgn_game.cpp
@@ -1,6 +1,6 @@
-#include "pgn_game.hpp"
-#include "read_pgn_data.hpp"
-#include "master_tablebase.hpp"
+#include "process_pgn/pgn_game.hpp"
+#include "process_pgn/read_pgn_data.hpp"
+#include "tablebase/master_tablebase.hpp"
 
 bool PgnGame::read_metadata_line(std::string &line)
 {
