+++ src/tablebase/tablebase.cpp
@@ -1,4 +1,4 @@
-#include "tablebase.hpp"
+#include "tablebase/tablebase.hpp"
 #include "util.hpp"
 #include <set>
 
@@ -7,9 +7,18 @@ long tablebase_position_hash_hash_distribution[64] = {0};
 
 std::mutex tablebase_position_hash_distribution_mutex;
 
-bool compare_move_edge(MoveEdge move_edge1, MoveEdge move_edge2)
+/* 
+  This function calculates and returns the move key, which is a concatenation 
+  of the source square, destination square, and promotion piece (complete 
+  information necessary to understand a move). For castling moves, even though
+  two pieces move, the src_square and dest_square pertain only to the king.
+*/
+uint32_t generate_move_key(uint8_t src_square, uint8_t dest_square, uint8_t promotion_piece)
 {
-    return move_edge1.m_times_played < move_edge2.m_times_played;
+    // move key is bit-wise concatenation of
+    // (empty/reserved) + start_square + end_square + promotion_piece
+    // 8 bits             8 bits         8 bits       8 bits
+    return (src_square << 16) + (dest_square << 8) + promotion_piece;
 }
 
 bool compare_key_move_pair(std::pair<MoveKey, MoveEdge> p1, std::pair<MoveKey, MoveEdge> p2)
